% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{busy-wellbeing}
\alias{busy-wellbeing}
\title{Social Pressure to be Busy and Emotional Wellbeing}
\format{
A data frame with variables:
\describe{
\item{participantID}{categorical. Participant ID}
\item{day}{ordinal. Day (Ranges from 1 to 8)}
\item{modifiedBusy}{continuous. Daily perceived social pressure to be busy (3-item)}
\item{modifiedBusy_mn}{continuous. Person mean of modifiedBusy (between-person component)}
\item{modifiedBusy_dev}{continuous. Person mean centred modifiedBusy (within-person component)}
\item{fouritemBusy}{continuous. Daily perceived social pressure to be busy (4-item)}
\item{fouritemBusy_mn}{continuous. Person mean of fouritemBusy (between-person component)}
\item{fouritemBusy_dev}{continuous. Person mean centred fouritemBusy (within-person component)}
\item{dayPA}{continuous. Daily positive affect}
\item{dayNA}{continuous. Daily negative affect}
\item{dayAnx}{continuous. Daily anxiety}
\item{dayDep}{continuous. Daily depressive symptoms}
\item{demogAge}{continuous. Age in years}
\item{demogAge.c}{continuous. Age in years (Grand mean centered)}
\item{demogSex}{binary. Gender (1 = Male, 0 = Female)}
\item{demogRace}{binary. Race (0 = Chinese, 1 = non-Chinese)}
\item{sesIncomeHousehold}{ordinal. Monthly household income}
\item{sesIncomeHousehold.c}{continuous. Monthly household income (Grand mean centered)}
\item{sesLadderCommunity}{ordinal. Subjective SES in community}
\item{sesLadderCommunity.c}{continuous. Subjective SES in community (Grand mean centered)}
\item{b5a}{continuous. Big 5 Agreeableness}
\item{b5c}{continuous. Big 5 Conscientiousness}
\item{b5e}{continuous. Big 5 Extraversion}
\item{b5n}{continuous. Big 5 Neuroticism}
\item{b5o}{continuous. Big 5 Open-mindedness}
\item{b5a.c}{continuous. Big 5 Agreeableness (Grand mean centered)}
\item{b5c.c}{continuous. Big 5 Conscientiousness (Grand mean centered)}
\item{b5e.c}{continuous. Big 5 Extraversion (Grand mean centered)}
\item{b5n.c}{continuous. Big 5 Neuroticism (Grand mean centered)}
\item{b5o.c}{continuous. Big 5 Open-mindedness (Grand mean centered)}
\item{emoComplexAcceptPersonal}{continuous. Acceptance of mixed emotions personal expectancies subscale}
\item{emoComplexAcceptPersonal.c}{continuous. Acceptance of mixed emotions personal expectancies subscale (Grand mean centered)}
\item{emoComplexAcceptSocial}{continuous. Acceptance of mixed emotions social expectancies subscale}
\item{emoComplexAcceptSocial.c}{continuous. Acceptance of mixed emotions social expectancies subscale (Grand mean centered)}
\item{emoComplexAcceptOverall}{continuous. Acceptance of mixed emotions}
\item{emoComplexAcceptOverall.c}{continuous. Acceptance of mixed emotions (Grand mean centered)}
\item{SEDAS_B_1}{ordinal. [ITEM] Daily perceived social pressure to be busy: Today, I felt a great deal of pressure from others to always be busy doing something productive}
\item{SEDAS_B_2}{ordinal. [ITEM] Daily perceived social pressure to be busy: Today, people expected me to always be busy doing something productive}
\item{SEDAS_B_3_REVERSED}{ordinal. [ITEM] Daily perceived social pressure to be busy: People valued me today, even though I was not always busy doing something productive (Reversed: 8 - original score)}
\item{SEDAS_B_4}{ordinal. [ITEM] Daily perceived social pressure to be busy: People were disapproving of me not always being busy doing something productive}
\item{NA_H_1}{ordinal. [ITEM] Negative affect: Angry}
\item{NA_H_2}{ordinal. [ITEM] Negative affect: Hostile}
\item{NA_H_3}{ordinal. [ITEM] Negative affect: Irritable}
\item{NA_M_4}{ordinal. [ITEM] Negative affect: Nervous}
\item{NA_M_5}{ordinal. [ITEM] Negative affect: Anxious}
\item{NA_M_6}{ordinal. [ITEM] Negative affect: Tense}
\item{NA_L_7}{ordinal. [ITEM] Negative affect: Dejected}
\item{NA_L_8}{ordinal. [ITEM] Negative affect: Sad}
\item{NA_L_9}{ordinal. [ITEM] Negative affect: Unhappy}
\item{PA_H_1}{ordinal. [ITEM] Positive affect: Energetic}
\item{PA_H_2}{ordinal. [ITEM] Positive affect: Enthusiastic}
\item{PA_H_3}{ordinal. [ITEM] Positive affect: Excited}
\item{PA_M_4}{ordinal. [ITEM] Positive affect: Happy}
\item{PA_M_5}{ordinal. [ITEM] Positive affect: Cheerful}
\item{PA_M_6}{ordinal. [ITEM] Positive affect: Pleasant}
\item{PA_L_7}{ordinal. [ITEM] Positive affect: Calm}
\item{PA_L_8}{ordinal. [ITEM] Positive affect: Content}
\item{PA_L_9}{ordinal. [ITEM] Positive affect: Relaxed}
\item{STAI_S_1}{ordinal. [ITEM] Anxiety: I feel upset today}
\item{STAI_S_2}{ordinal. [ITEM] Anxiety: I feel anxious today}
\item{STAI_S_3}{ordinal. [ITEM] Anxiety: I feel nervous today}
\item{STAI_S_4}{ordinal. [ITEM] Anxiety: I am jittery today}
\item{STAI_S_5}{ordinal. [ITEM] Anxiety: I feel over-excited and rattled today}
\item{CES_D_R_1}{ordinal. [ITEM] Depression: I was bothered by things that usually don't bother me.}
\item{CES_D_R_2}{ordinal. [ITEM] Depression: I had trouble keeping my mind on what I was doing.}
\item{CES_D_R_3}{ordinal. [ITEM] Depression: I felt depressed.}
\item{CES_D_R_4}{ordinal. [ITEM] Depression: I felt that everything I did was an effort.}
\item{CES_D_R_5}{ordinal. [ITEM] Depression: I felt hopeful about the future. (Reversed)}
\item{CES_D_R_6}{ordinal. [ITEM] Depression: I felt fearful.}
\item{CES_D_R_7}{ordinal. [ITEM] Depression: My sleep was restless.}
\item{CES_D_R_8}{ordinal. [ITEM] Depression: I was happy. (Reversed)}
\item{CES_D_R_9}{ordinal. [ITEM] Depression: I felt lonely.}
\item{CES_D_R_10}{ordinal. [ITEM] Depression: I could not get going.}
\item{ECS_03}{ordinal. [ITEM] ECA: It is very important to me not to feel mixed emotions. (Reversed)}
\item{ECS_04}{ordinal. [ITEM] ECA: I would always try to avoid feeling mixed emotions. (Reversed)}
\item{ECS_05}{ordinal. [ITEM] ECA: I shouldn't feel mixed emotions. (Reversed)}
\item{ECS_06}{ordinal. [ITEM] ECA: Feeling mixed emotions is normal.}
\item{ECS_07}{ordinal. [ITEM] ECA: Feeling mixed emotions is an important part of life.}
\item{ECS_08}{ordinal. [ITEM] ECA: It is important that others don't see me as a person with mixed emotions. (Reversed)}
\item{ECS_09}{ordinal. [ITEM] ECA: Other people expect me not to feel mixed emotions. (Reversed)}
\item{ECS_10}{ordinal. [ITEM] ECA: People like me less when I feel mixed emotions. (Reversed)}
\item{ECS_11}{ordinal. [ITEM] ECA: I think society accepts people who feel mixed emotions.}
\item{ECS_12}{ordinal. [ITEM] ECA: I think it is socially acceptable to feel mixed emotions.}
\item{ECS_13}{ordinal. [ITEM] ECA: Society generally expects people not to feel mixed emotions. (Reversed)}
\item{BFI_E_01}{ordinal. [ITEM] B5E: Tends to be quiet.}
\item{BFI_A_02}{ordinal. [ITEM] B5A: Is compassionate, has a soft heart.}
\item{BFI_C_03}{ordinal. [ITEM] B5C: Tends to be disorganized.}
\item{BFI_N_04}{ordinal. [ITEM] B5N: Worries a lot.}
\item{BFI_O_05}{ordinal. [ITEM] B5O: Is fascinated by art, music, or literature.}
\item{BFI_E_06}{ordinal. [ITEM] B5E: Is dominant, acts as a leader.}
\item{BFI_A_07}{ordinal. [ITEM] B5A: Is sometimes rude to others.}
\item{BFI_C_08}{ordinal. [ITEM] B5C: Has difficulty getting started on tasks.}
\item{BFI_N_09}{ordinal. [ITEM] B5N: Tends to feel depressed, blue.}
\item{BFI_O_10}{ordinal. [ITEM] B5O: Has little interest in abstract ideas.}
\item{BFI_E_11}{ordinal. [ITEM] B5E: Is full of energy.}
\item{BFI_A_12}{ordinal. [ITEM] B5A: Assumes the best about people.}
\item{BFI_C_13}{ordinal. [ITEM] B5C: Is reliable, can always be counted on.}
\item{BFI_N_14}{ordinal. [ITEM] B5N: Is emotionally stable, not easily upset.}
\item{BFI_O_15}{ordinal. [ITEM] B5O: Is original, comes up with new ideas.}
\item{BFI_E_16}{ordinal. [ITEM] B5E: Is outgoing, sociable.}
\item{BFI_A_17}{ordinal. [ITEM] B5A: Can be cold and uncaring.}
\item{BFI_C_18}{ordinal. [ITEM] B5C: Keeps things neat and tidy.}
\item{BFI_N_19}{ordinal. [ITEM] B5N: Is relaxed, handles stress well.}
\item{BFI_O_20}{ordinal. [ITEM] B5O: Has few artistic interests.}
\item{BFI_E_21}{ordinal. [ITEM] B5E: Prefers to have others take charge.}
\item{BFI_A_22}{ordinal. [ITEM] B5A: Is respectful, treats others with respect.}
\item{BFI_C_23}{ordinal. [ITEM] B5C: Is persistent, works until the task is finished.}
\item{BFI_N_24}{ordinal. [ITEM] B5N: Feels secure, comfortable with self.}
\item{BFI_O_25}{ordinal. [ITEM] B5O: Is complex, a deep thinker.}
\item{BFI_E_26}{ordinal. [ITEM] B5E: Is less active than other people.}
\item{BFI_A_27}{ordinal. [ITEM] B5A: Tends to find fault with others.}
\item{BFI_C_28}{ordinal. [ITEM] B5C: Can be somewhat careless.}
\item{BFI_N_29}{ordinal. [ITEM] B5N: Is temperamental, gets emotional easily.}
\item{BFI_O_30}{ordinal. [ITEM] B5O: Has little creativity.}
}
}
\source{
Lua VYQ, Majeed NM, Leung AK, Hartanto A. (2022). A daily within-person investigation on the link between social expectancies to be busy and emotional wellbeing: the moderating role of emotional complexity acceptance. Cognition and Emotion, 1-8. https://doi.org/10.1080/02699931.2022.2054778
}
\description{
A daily diary study examining how social pressure to appear busy relates to emotional wellbeing. Data were collected from 261 participants over approximately 7 days each, yielding 1,827 person-day observations. Participants reported daily busyness, positive affect, negative affect, anxiety, and depressive symptoms, alongside person-level demographics, socioeconomic status, Big Five personality traits, and emotional complexity measures. Ideal for teaching multilevel modeling, within-person centering, and longitudinal data analysis.
}
