% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{discrimination-cognitive}
\alias{discrimination-cognitive}
\title{Everyday Discrimination and Cognitive Failures (Singapore Daily Diary Study)}
\format{
A data frame with variables:
\describe{
\item{participantID}{categorical. Unique participant identifier}
\item{day}{ordinal. Study day number (1 to 7)}
\item{age}{continuous. Participant age in years}
\item{age.c10}{continuous. Age centered at 10-year intervals}
\item{female}{binary. Gender (1 = female, 0 = male)}
\item{female.f}{categorical. Gender as factor (YES = female, NO = male)}
\item{nonChinese}{binary. Ethnicity (0 = Chinese, 1 = non-Chinese)}
\item{nonChinese.f}{categorical. Ethnicity as factor (YES = non-Chinese, NO = Chinese)}
\item{sesInc}{ordinal. Household income bracket (1-6; higher values reflect higher household income)}
\item{sesInc.c}{continuous. Grand mean centred household income}
\item{sesLad}{ordinal. Subjective socioeconomic status in community (1 = lowest, 10 = highest)}
\item{sesLad.c}{continuous. Grand mean centred community ladder}
\item{discrimination}{continuous. Daily discrimination composite score (mean of 9 items)}
\item{itemDisc1}{ordinal. [ITEM] Discrimination (1=never, 4=often): You were treated with less courtesy or respect than other people}
\item{itemDisc2}{ordinal. [ITEM] Discrimination (1=never, 4=often): You received poorer service than other people at restaurants or stores}
\item{itemDisc3}{ordinal. [ITEM] Discrimination (1=never, 4=often): People acted as if they think you are not smart}
\item{itemDisc4}{ordinal. [ITEM] Discrimination (1=never, 4=often): People acted as if they are afraid of you}
\item{itemDisc5}{ordinal. [ITEM] Discrimination (1=never, 4=often): People acted as if they think you are dishonest}
\item{itemDisc6}{ordinal. [ITEM] Discrimination (1=never, 4=often): People acted as if they are better than you}
\item{itemDisc7}{ordinal. [ITEM] Discrimination (1=never, 4=often): You were called names or insulted}
\item{itemDisc8}{ordinal. [ITEM] Discrimination (1=never, 4=often): You were threatened or harassed}
\item{itemDisc9}{ordinal. [ITEM] Discrimination (1=never, 4=often): You were followed around in stores}
\item{dayCogFail}{continuous. Daily cognitive failures composite score (mean of 13 items)}
\item{dayCogFail.m}{continuous. Person mean of dayCogFail (between-person component)}
\item{itemDayCogFail01}{count. [ITEM] Cognitive failure (0=never, 3=many times): Did not notice signposts when driving/walking}
\item{itemDayCogFail02}{count. [ITEM] Cognitive failure (0=never, 3=many times): Forgot the purpose of going somewhere}
\item{itemDayCogFail03}{count. [ITEM] Cognitive failure (0=never, 3=many times): Failed to notice something pointed out to you}
\item{itemDayCogFail04}{count. [ITEM] Cognitive failure (0=never, 3=many times): Accidentally threw away something you meant to keep}
\item{itemDayCogFail05}{count. [ITEM] Cognitive failure (0=never, 3=many times): Daydreamed when you should have been concentrating}
\item{itemDayCogFail06}{count. [ITEM] Cognitive failure (0=never, 3=many times): Forgot whether you had done something}
\item{itemDayCogFail07}{count. [ITEM] Cognitive failure (0=never, 3=many times): Failed to read information properly before proceeding}
\item{itemDayCogFail08}{count. [ITEM] Cognitive failure (0=never, 3=many times): Said something and later wondered if it made sense}
\item{itemDayCogFail09}{count. [ITEM] Cognitive failure (0=never, 3=many times): Failed to take notice of something on first reading}
\item{itemDayCogFail10}{count. [ITEM] Cognitive failure (0=never, 3=many times): Lost track of what you were saying mid-sentence}
\item{itemDayCogFail11}{count. [ITEM] Cognitive failure (0=never, 3=many times): Forgot an important appointment}
\item{itemDayCogFail12}{count. [ITEM] Cognitive failure (0=never, 3=many times): Had trouble concentrating on a task}
\item{itemDayCogFail13}{count. [ITEM] Cognitive failure (0=never, 3=many times): Forgot where you had put something}
\item{dayStressorExp}{binary. Daily stressor exposure (1 = experienced at least one stressor, 0 = none)}
\item{dayStressorExp.m}{continuous. Person mean of dayStressorExp (between-person component)}
\item{dayStressorExp.f}{categorical. Daily stressor exposure as factor (YES/NO)}
\item{itemDayStressorArgue}{binary. [ITEM] Stressor (0=no, 1=yes): Had an argument or disagreement with someone}
\item{itemDayStressorAvoid}{binary. [ITEM] Stressor (0=no, 1=yes): Avoided someone to prevent an argument}
\item{itemDayStressorWork}{binary. [ITEM] Stressor (0=no, 1=yes): Had a work- or school-related stressor}
\item{itemDayStressorHome}{binary. [ITEM] Stressor (0=no, 1=yes): Had a home- or family-related stressor}
\item{itemDayStressorDiscrimination}{binary. [ITEM] Stressor (0=no, 1=yes): Experienced discrimination or prejudice}
\item{itemDayStressorNetwork}{binary. [ITEM] Stressor (0=no, 1=yes): Had a problem with a friend or member of your social network}
\item{itemDayStressorOther}{binary. [ITEM] Stressor (0=no, 1=yes): Experienced another stressor not listed above}
}
}
\source{
Majeed NM, Kasturiratna KS, Lua VY, Li MY, Hartanto A. (2023). Discrimination and cognitive failures in Singapore and the US: An investigation of between- and within-persons associations through multilevel modelling. International Journal of Intercultural Relations, 94, 101805. https://doi.org/10.1016/j.ijintrel.2023.101805
}
\description{
A seven-day daily diary study examining the association between everyday discrimination and cognitive failures among 484 undergraduate students at Singapore universities (Study 1 from Majeed et al., 2023). Participants completed nightly questionnaires reporting daily discrimination experiences, cognitive failures, and daily stressor exposure, alongside person-level demographics and socioeconomic variables. Multilevel modelling revealed that experiencing discrimination was associated with poorer cognitive functioning at both within- and between-person levels, even after controlling for demographic covariates and daily stressor exposure. Ideal for teaching multilevel modelling, within-person centering, and the study of discrimination and cognitive health in a daily diary context.
}
