% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kenya_accidents_2017}
\alias{kenya_accidents_2017}
\title{Kenya road accident database, 2017}
\format{
A data frame with 740 rows and 15 columns:
\describe{
  \item{TIME 24 HOURS}{Time of incident on a 24-hour clock}
  \item{BASE/SUB BASE}{Police base or sub-base responsible}
  \item{COUNTY}{County in which the incident occurred}
  \item{ROAD}{Road on which the incident occurred}
  \item{PLACE}{Specific location of the incident}
  \item{MV INVOLVED}{Information on the motor vehicle involved in the incident}
  \item{BRIEF ACCIDENT DETAILS}{Free-text description of the incident}
  \item{NAME OF VICTIM}{Free-text description of victim name}
  \item{GENDER}{Gender of victim(s)}
  \item{AGE}{Age of victim(s)}
  \item{CAUSE CODE}{Numeric cause code}
  \item{VICTIM}{Victim category}
  \item{NO.}{Number of victims}
  \item{Date DD/MM/YYYY}{Date of incident}
  \item{...15}{Extra Column for illustration purposes}
}
}
\source{
Humanitarian Data Exchange.
  \url{https://data.humdata.org/dataset/kenya-road-accidents-database}
}
\usage{
kenya_accidents_2017
}
\description{
Fatal road accident records from Kenya, covering 2017.
Each row is one incident.
}
\keyword{datasets}
