% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kenya_flamingos}
\alias{kenya_flamingos}
\title{Flamingo measurements from Kenya's Rift Valley}
\format{
A data frame with 344 rows and 8 columns:
\describe{
  \item{lake}{Lake at which the bird was recorded: Bogoria,
    Elementaita, or Nakuru}
  \item{species}{Greater Flamingo or Lesser Flamingo}
  \item{bill_length_mm}{Bill length, in millimetres}
  \item{bill_depth_mm}{Bill depth, in millimetres}
  \item{leg_length_mm}{Leg length, in millimetres}
  \item{body_mass_g}{Body mass, in grams}
  \item{sex}{Recorded sex of the bird}
  \item{year}{Year of measurement}
}
}
\source{
Collected as part of the IDEMS statistics textbook project.
}
\usage{
kenya_flamingos
}
\description{
Morphological measurements recorded from Greater and Lesser
Flamingos at three lakes in Kenya's Rift Valley between 2021
and 2023.
}
\keyword{datasets}
