% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset by its ID}
\usage{
load_dataset(id)
}
\arguments{
\item{id}{The dataset's ID, e.g. "titanic-survival"}
}
\value{
A data frame containing the dataset
}
\description{
Load a dataset by its ID
}
