% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_helpers.R
\name{search_by_concept}
\alias{search_by_concept}
\title{Search datasets by teaching concept}
\usage{
search_by_concept(concept)
}
\arguments{
\item{concept}{Teaching concept (e.g. "confounding", "missing_data")}
}
\value{
A data frame of matching datasets
}
\description{
Search datasets by teaching concept
}
