% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_helpers.R
\name{search_by_method}
\alias{search_by_method}
\title{Search datasets by statistical method}
\usage{
search_by_method(method)
}
\arguments{
\item{method}{Statistical method (e.g. "logistic_regression", "anova")}
}
\value{
A data frame of matching datasets
}
\description{
Search datasets by statistical method
}
