% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_helpers.R
\name{search_by_variable}
\alias{search_by_variable}
\title{Search datasets by variable type}
\usage{
search_by_variable(variable_type)
}
\arguments{
\item{variable_type}{Variable type (e.g. "continuous", "binary", "count")}
}
\value{
A data frame of matching datasets
}
\description{
Search datasets by variable type
}
